+++ mb/models/Icosahedron.cpp
@@ -20,22 +20,54 @@
  *
  */
 
-#ifndef __MB_POLYHEDRON__
-#define __MB_POLYHEDRON__
-
-#include <mb/api.h>
-
-#include "Drawable.hpp"
+#include "Icosahedron.hpp"
+#include <cmath>
 
 namespace MB
 {
-    class Polyhedron: public Drawable
-    {
-    protected:
-		Polyhedron();
-        void create(std::vector<float>& verts, std::vector<unsigned int>& el, 
-			const float& radius, const unsigned int& subdivisions);
-    };
-}
+	Icosahedron::Icosahedron(const float& radius, const unsigned int& subdivisions)
+	: Polyhedron()
+	{
+		float t = (1.0f + std::sqrt(5.0f)) / 2.0f;
 
-#endif /* __MB_POLYHEDRON__ */
+		std::vector<float> verts = {
+			-1.0f,  t,  0.0f,
+			 1.0f,  t,  0.0f,
+			-1.0f, -t,  0.0f,
+			 1.0f, -t,  0.0f,
+
+			 0.0f, -1.0f,  t,
+			 0.0f,  1.0f,  t,
+			 0.0f, -1.0f, -t,
+			 0.0f,  1.0f, -t,
+
+			 t,  0.0f, -1.0f,
+			 t,  0.0f,  1.0f,
+			-t,  0.0f, -1.0f,
+			-t,  0.0f,  1.0f
+		};
+		std::vector<unsigned int> el = {
+			 0, 11,   5,
+			 0,  5,   1,
+			 0,  1,   7,
+			 0,  7,  10,
+			 0, 10,  11,
+			 1,  5,   9,
+			 5, 11,   4,
+			11, 10,   2,
+			10,  7,   6,
+			 7,  1,   8,
+			 3,  9,   4,
+			 3,  4,   2,
+			 3,  2,   6,
+			 3,  6,   8,
+			 3,  8,   9,
+			 4,  9,   5,
+			 2,  4,  11,
+			 6,  2,  10,
+			 8,  6,   7,
+			 9,  8,   1
+		};
+		this->create(verts, el, radius, subdivisions);
+	}
+}